some complex header refactors
