--- packages/client/src/styles/logged-in-menu.module.css
@@ -1,91 +1,0 @@
-.notificationsDropdown {
-  padding-inline: var(--edge-spacing);
-  max-width: 570px;
-  background-color: var(--a-surface-subtle);
-  padding-block: var(--a-spacing-8);
-}
-
-.myPageMenuDropdown {
-  padding-block: var(--a-spacing-2) var(--a-spacing-4);
-  padding-inline: var(--edge-spacing);
-  max-width: 1024px;
-}
-
-.myPageMenuButton.myPageMenuButton {
-  display: none;
-}
-
-@media (min-width: 768px) {
-  .myPageMenuButton.myPageMenuButton {
-    display: flex;
-  }
-}
-
-.loggedInMenu {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  padding: 0 var(--a-spacing-2);
-}
-
-@media (min-width: 768px) {
-  .loggedInMenu {
-    background-color: rgba(230, 240, 255, 1);
-  }
-}
-
-.myPageMenuHeading {
-  font-size: var(--a-font-size-medium);
-  font-weight: var(--a-font-weight-bold);
-}
-
-.link {
-  color: var(--a-text-action);
-  text-decoration-line: underline;
-  padding-top: 0.5rem /* 8px */;
-  padding-bottom: 0.75rem /* 12px */;
-  display: inline-block;
-  text-underline-offset: 3px;
-}
-
-.notificationsUnread {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  background-color: var(--a-icon-danger);
-  position: absolute;
-  bottom: 3px;
-  right: -4px;
-  display: none;
-}
-
-.notificationsUnread {
-  display: block;
-}
-
-.notificationsIconWrapper {
-  position: relative;
-  display: flex;
-}
-
-[aria-expanded='true'] .notificationsIcon {
-  display: none;
-}
-
-.closeIcon {
-  display: none;
-}
-
-[aria-expanded='true'] .closeIcon {
-  display: block;
-}
-
-.logoutButton.logoutButton {
-  display: none;
-}
-
-@media (min-width: 768px) {
-  .logoutButton.logoutButton {
-    display: flex;
-  }
-}
