+++ packages/server/src/views/header/complex-header.ts
@@ -1,18 +1,22 @@
 import clsx from 'clsx';
+import menuCls from 'decorator-client/src/styles/complex-header-menu.module.css';
+import cls from 'decorator-client/src/styles/header.module.css';
+import menuItemsCls from 'decorator-client/src/styles/menu-items.module.css';
+import opsMessagesCls from 'decorator-client/src/styles/ops-messages.module.css';
+import { ContextLink } from 'decorator-shared/context';
 import html, { Template } from 'decorator-shared/html';
 import { Context, Language } from 'decorator-shared/params';
-import {
-  LinkGroup,
-  MainMenuContextLink,
-  Node,
-  Texts,
-} from 'decorator-shared/types';
-import { ComplexHeaderNavbarItems } from './navbar-items/complex-header-navbar-items';
-import { ContextLink } from 'decorator-shared/context';
-import { LenkeMedSporing } from 'decorator-shared/views/lenke-med-sporing-helpers';
-import cls from 'decorator-client/src/styles/header.module.css';
+import { LinkGroup, MainMenuContextLink, Texts } from 'decorator-shared/types';
 import utilsCls from 'decorator-shared/utilities.module.css';
-import opsMessagesCls from 'decorator-client/src/styles/ops-messages.module.css';
+import {
+  BurgerIcon,
+  LoginIcon,
+  SearchIcon,
+} from 'decorator-shared/views/icons';
+import { DropdownMenu } from '../dropdown-menu';
+import { IconButton } from '../icon-button';
+import { SearchForm } from '../search-form';
+import { MainMenu } from './main-menu';
 
 const frontPageUrl = (context: Context, language: Language) => {
   if (language === 'en') {
@@ -31,11 +35,8 @@ const frontPageUrl = (context: Context, language: Language) => {
 
 export type ComplexHeaderProps = {
   texts: Texts;
-  innlogget: boolean;
   decoratorUtilsContainer?: Template;
-  myPageMenu?: Node[];
   context: Context;
-  name?: string;
   language: Language;
   mainMenuLinks: LinkGroup[];
   mainMenuContextLinks: MainMenuContextLink[];
@@ -48,29 +49,25 @@ export function ComplexHeader({
   mainMenuLinks,
   mainMenuContextLinks,
   texts,
-  innlogget,
   context: currentContext,
-  myPageMenu,
-  name,
   decoratorUtilsContainer,
 }: ComplexHeaderProps) {
   return html`
     <header class="${cls.siteheader}">
       <div class="${cls.hovedmenyWrapper} ${utilsCls.contentContainer}">
         <div class="${cls.hovedmenyContent}">
-          ${LenkeMedSporing({
-            className: cls.logo,
-            href: '/',
-            analyticsEventArgs: {
+          <a
+            is="lenke-med-sporing"
+            href="/"
+            class="${cls.logo}"
+            data-analytics-event-args="${JSON.stringify({
               category: 'dekorator-header',
               action: 'navlogo',
-            },
-            dataAttachContext: true,
-            children: html`<img
-              src="/public/ikoner/meny/nav-logo-red.svg"
-              alt="NAV"
-            />`,
-          })}
+            })}"
+            data-attach-context
+          >
+            <img src="/public/ikoner/meny/nav-logo-red.svg" lt="NAV" />
+          </a>
           <div class="${cls.arbeidsflate}">
             ${language === 'nb' &&
             contextLinks?.map(
@@ -94,19 +91,58 @@ export function ComplexHeader({
             )}
           </div>
         </div>
-        ${ComplexHeaderNavbarItems({
-          mainMenuTitle:
-            currentContext === 'privatperson'
-              ? texts.how_can_we_help
-              : texts[`rolle_${currentContext}`],
-          frontPageUrl: frontPageUrl(currentContext, language),
-          innlogget,
-          texts,
-          myPageMenu: myPageMenu as Node[],
-          mainMenuLinks,
-          contextLinks: mainMenuContextLinks,
-          name,
-        })}
+        <div class="${menuItemsCls.menuItems}">
+          <user-menu>
+            ${IconButton({
+              id: 'login-button',
+              Icon: LoginIcon({}),
+              text: texts.login,
+            })}
+          </user-menu>
+          <div class="${menuItemsCls.menuItemsUniversalLinks}">
+            ${DropdownMenu({
+              button: IconButton({
+                Icon: BurgerIcon(),
+                text: texts.menu,
+              }),
+              dropdownContent: html`
+                <div class="${menuCls.menuContent}">
+                  <search-menu class="${menuCls.searchMenu}">
+                    ${SearchForm({ texts })}
+                  </search-menu>
+                  <main-menu>
+                    ${MainMenu({
+                      title:
+                        currentContext === 'privatperson'
+                          ? texts.how_can_we_help
+                          : texts[`rolle_${currentContext}`],
+
+                      frontPageUrl: frontPageUrl(currentContext, language),
+                      links: mainMenuLinks,
+                      contextLinks: mainMenuContextLinks,
+                      texts,
+                    })}
+                  </main-menu>
+                </div>
+              `,
+            })}
+            ${DropdownMenu({
+              button: IconButton({
+                Icon: SearchIcon({
+                  menuSearch: true,
+                }),
+                text: texts.search,
+                className: menuItemsCls.searchButton,
+              }),
+              dropdownClass: menuItemsCls.searchDropdown,
+              dropdownContent: html`
+                <search-menu class="${menuItemsCls.searchMenu}" data-auto-focus>
+                  ${SearchForm({ texts })}
+                </search-menu>
+              `,
+            })}
+          </div>
+        </div>
       </div>
     </header>
     <ops-messages class="${opsMessagesCls.opsMessages}"></ops-messages>
